fix serialization of local_operators. (#93)
Left the comment for reference to old implementation that is broken. To
remove.
We can add a test. op.acting on is consistently returning a list and its
serialization works every time. op._acting_on doesnt always exist and
when it does its behaviour changes depending on when its called. The
current serialization seems to be stable changing from LocalOperators
and LocalJaxOperators.

GitOrigin-RevId: b2b9e80335f491a0d5020a365f866f31e7d89a6a